UI splitting experimens: EntriesView and QueryView
