+++ src/lib/systemd_journal/entries_dialog.rb
@@ -18,12 +18,9 @@
 
 require "yast"
 require "ui/dialog"
+require "systemd_journal/entries_view"
 require "systemd_journal/query_presenter"
 require "systemd_journal/query_dialog"
-
-Yast.import "UI"
-Yast.import "Label"
-Yast.import "Popup"
 
 module SystemdJournal
   # Dialog to display journal entries with several filtering options
@@ -33,109 +30,38 @@ module SystemdJournal
       textdomain "systemd_journal"
 
       @query = QueryPresenter.new
-      @search = ""
-      read_journal_entries
+      execute_query
+      @view = EntriesView.new(@query)
     end
 
-    # Main dialog layout
     def dialog_content
-      VBox(
-        # Header
-        Heading(_("Journal entries")),
-        # Filters
-        Left(
-          HBox(
-            Label(_("Displaying entries with the following text")),
-            HSpacing(1),
-            InputField(Id(:search), Opt(:hstretch, :notify), "", @search)
-          )
-        ),
-        ReplacePoint(Id(:query), query_description),
-        VSpacing(0.3),
-        # Log entries
-        table,
-        VSpacing(0.3),
-        # Footer buttons
-        footer
-      )
+      @view.content
     end
 
-    # Dialog options
     def dialog_options
-      Opt(:decorated, :defaultsize)
+      @view.dialog_options
     end
 
     # Event callback for the 'change filter' button.
     def filter_handler
       return unless read_query
-      read_journal_entries
-      redraw_query
-      redraw_table
+      @view.redraw_query
+      execute_query
+      @view.redraw_table
     end
 
     # Event callback for change in the content of the search box
     def search_handler
-      read_search
-      redraw_table
+      @view.redraw_table
     end
 
     # Event callback for the 'refresh' button
     def refresh_handler
-      read_journal_entries
-      redraw_table
+      execute_query
+      @view.redraw_table
     end
 
     private
-
-    # Table widget (plus wrappers) to display log entries
-    def table
-      headers = @query.columns.map { |c| c[:label] }
-
-      Table(
-        Id(:table),
-        Opt(:keepSorting),
-        Header(*headers),
-        table_items
-      )
-    end
-
-    def table_items
-      # Reduce it to an array with only the visible fields
-      entries_fields = @journal_entries.map do |entry|
-        @query.columns.map { |c| entry.send(c[:method]) }
-      end
-      # Grep for entries matching @search in any visible field
-      entries_fields.select! do |fields|
-        fields.any? { |f| Regexp.new(@search, Regexp::IGNORECASE).match(f) }
-      end
-      # Return the result as an array of Items
-      entries_fields.map { |fields| Item(*fields) }
-    end
-
-    def footer
-      HBox(
-        HWeight(1, PushButton(Id(:filter), _("Change filter..."))),
-        HStretch(),
-        HWeight(1, PushButton(Id(:refresh), _("Refresh"))),
-        HStretch(),
-        HWeight(1, PushButton(Id(:cancel), Yast::Label.QuitButton))
-      )
-    end
-
-    def query_description
-      VBox(
-        Left(Label(" - #{@query.interval_description}")),
-        Left(Label(" - #{@query.filters_description}"))
-      )
-    end
-
-    def redraw_query
-      Yast::UI.ReplaceWidget(Id(:query), query_description)
-    end
-
-    def redraw_table
-      Yast::UI.ChangeWidget(Id(:table), :Items, table_items)
-    end
 
     # Asks the user the new query options using SystemdJournal::QueryDialog.
     #
@@ -146,6 +72,7 @@ module SystemdJournal
       query = QueryDialog.new(@query).run
       if query
         @query = query
+        @view.query = query
         log.info "New query is #{@query}."
         true
       else
@@ -154,21 +81,13 @@ module SystemdJournal
       end
     end
 
-    # Gets the new search string from the interface
-    def read_search
-      @search = Yast::UI.QueryWidget(Id(:search), :Value)
-      log.info "Search string set to '#{@search}'"
-    end
-
     # Reads the journal entries from the system
-    def read_journal_entries
-      log.info "Calling journalctl with #{@query.journalctl_options} "\
-        "and #{@query.journalctl_matches}"
-      @journal_entries = @query.entries
-      log.info "Call to journalctl returned #{@journal_entries.size} entries."
+    def execute_query
+      log.info "Executing query #{@query}"
+      @query.execute
+      log.info "Call to journalctl returned #{@query.entries.size} entries."
     rescue => e
       log.warn e.message
-      @journal_entries = []
       Yast::Popup.Message(e.message)
     end
   end
