+++ src/lib/systemd_journal/query.rb
@@ -31,6 +31,7 @@ module SystemdJournal
     # @return [Array] matches in the format expected by Journalctl
     # @see SystemdJournal::Journalctl#initialize
     attr_reader :journalctl_matches
+    attr_reader :entries
 
     # Creates a new query based on the time interval and some additional filters
     #
@@ -58,17 +59,22 @@ module SystemdJournal
       else
         @journalctl_matches = [filters["match"]].flatten
       end
-
       calculate_options
+
+      @entries = []
     end
 
-    # Calls journalctl and returns an Array of Entry objects
-    def entries
-      Entry.all(options: journalctl_options, matches: journalctl_matches)
+    # Calls journalctl and XXX Array of Entry objects
+    def execute
+      @entries = Entry.all(
+        options: journalctl_options,
+        matches: journalctl_matches
+      )
     end
 
     def to_s
-      "<interval: #{@interval}, filters: #{@filters}>"
+      "<interval: #{@interval}, filters: #{@filters}, journalctl_options: "\
+        "#{@journalctl_options}, journalctl_matches #{@journalctl_matches}>"
     end
 
     # Array of system's boots registered in the journal.
